fourtrees:game: Remove boilerplate
Remove code from fourtrees#Game that was generic and did not have to do
with the actual four trees game. This primarily includes the code for
exiting, which is now managed by the scene director and a special
"Exit" scene.

In similar fashion remove code from main() that was specific to
fourtrees.

While I'm at it, I realized the scene director's rules pretty much have
to live in main() becaue they would contain references to all scenes.