fourtrees: Go between game and demo
Use the scene system to show the demo, let the user go into the game,
exit the game back to the demo and exit the demo to exit the program
alltogether.

As a result of this work, everything gained a Load function and I no
longer need to initialize stuff in Update(). I will probably need to
put in place some kind of resource cache to avoid unneccesarily
re-loading the same resources and ending up with multiple copies in
memory.

Will also need to consider unlaoding stuff to redoce the memory
footprint of inactive scenes.