fourtrees: Share most state between scenes
Use the GState structure to share most of the game state between scenes
so that we don't end up with multiple copies in memory. Also used it
along with the "Boiler" structure to share almost all of the drawing and
behavior code so that scenes end up only with code that is unique to
that scene. This enabled me to stop having "Demo" contain "Game", and
therefore stop displaying the game buttons when running "Demo"